fix: refactor-repo-dto-errorcode - #19
