+++ crossover-server-2/src/main/java/org/example/crossoverserver2/planeletter/dto/ResponseDto.java
@@ -1,4 +1,4 @@
-package org.example.crossoverserver2.planeletter.dto.response;
+package org.example.crossoverserver2.planeletter.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
