+++ crossover-server-2/src/main/java/org/example/crossoverserver2/planeletter/repository/UserRepository.java
@@ -3,7 +3,13 @@ package org.example.crossoverserver2.planeletter.repository;
 import org.example.crossoverserver2.planeletter.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.Optional;
 import java.util.UUID;
 
 public interface UserRepository extends JpaRepository<User, UUID> {
+        boolean existsByEmail(String email);
+        boolean existsByUserId(String userId);
+
+        Optional<User> findByUserId(String userId);
+
 }
